Modularize toolkit and decouple API backend
