+++ hallbayes/utils.py
